refactor(useMainSize): 반응형에 따른 리스트 사이즈 변경되는 훅 분리
