+++ src/components/pages/main/SearchResultSection.tsx
@@ -1,23 +1,15 @@
 'use client';
 import AllExperiencesList from '@/src/components/pages/main/AllExperiencesList';
 import { useActivities } from '@/src/hooks/pages/main/useActivities';
-import { useBreakPoint } from '@/src/hooks/useBreakPoint';
+import useMainSize from '@/src/hooks/pages/main/useMainSize';
 import { useSearchParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
-
-const MAX_SIZE = {
-  mo: 6,
-  tb: 4,
-  pc: 8,
-};
 
 export default function SearchResultSection() {
   const searchParams = useSearchParams();
   const currentSearch = searchParams.get('search');
-  const { isMd, isLg } = useBreakPoint();
   const [page, setPage] = useState(1);
-  const [size, setSize] = useState(MAX_SIZE['mo']);
-
+  const { size } = useMainSize();
   const { data: resultList, isLoading } = useActivities({
     page,
     size,
@@ -26,16 +18,13 @@ export default function SearchResultSection() {
   });
 
   useEffect(() => {
-    if (isLg && isMd) setSize(MAX_SIZE['pc']);
-    else if (isMd && !isLg) setSize(MAX_SIZE['tb']);
-    else setSize(MAX_SIZE['mo']);
+    const totalCount = resultList?.totalCount;
+    const lastPage = totalCount && Math.ceil(totalCount / size);
 
-    // 마지막 페이지네이션에서 pagePerView 가 변경되었을 때 페이지의 데이터가 빈값이면 이전 페이지네이션을 보여주도록 처리
-    // 조금 불안정. 다시 테스트 필요
-    if (resultList?.activities.length === 0) {
-      setPage((prev) => prev - 1);
+    if (lastPage && size > lastPage) {
+      setPage(lastPage);
     }
-  }, [isMd, isLg, resultList]);
+  }, [size, resultList]);
 
   useEffect(() => {
     // 검색값이 변경되면, 페이지네이션 초기화
